requirements.txt
gradio>=4.0.0
torch>=2.0.0
torchvision
Pillow>=9.0.0
numpy
opencv-python
ffmpeg-python
loguru
moviepy
einops
decord
peft
onnxruntime
pandas
matplotlib
sentencepiece
diffusers>=0.31.0
transformers>=4.52,<5
tokenizers>=0.20.3
accelerate>=1.1.1
tqdm
scipy>=1.14
imageio[ffmpeg]
easydict
ftfy
imageio-ffmpeg
sam2
